No longer use controller spec
