+++ app/models/micropost.rb
@@ -33,7 +33,7 @@ class Micropost < ApplicationRecord
   end
 
   def data
-    @data ||= YoutubeApi.get get_id(link)
+    @data ||= YoutubeApi.get_video_data(get_id_from_link(link))
   end
 
   def trim(text, lim)
@@ -42,33 +42,12 @@ class Micropost < ApplicationRecord
     text.slice(0, lim - 3) + '...'
   end
 
-  def get_id(link)
+  def get_id_from_link(link)
     return if link.nil?
 
     id = link.gsub(/(>|<)/i, '').split(/(vi\/|v=|\/v\/|youtu\.be\/|\/embed\/)/)
     return id if id[2].nil?
 
     id[2].split(/[^0-9a-z_\-]/i)[0]
-  end
-end
-
-
-# todo: separate file, service
-class YoutubeApi
-  BASE_URL = 'https://www.googleapis.com/youtube/v3/videos?part=id%2C+snippet'.freeze
-
-  AUTH_KEY = Figaro.env.youtube_api_key
-
-  def self.get(video_id)
-    options = { query: { id: video_id, key: AUTH_KEY } }
-    response = HTTParty.get(BASE_URL, options)
-    status = response.header.code
-    return {} if !status.eql?('200') || response['items'].blank?
-
-    {
-      id: video_id,
-      title: response['items'][0]['snippet']['title'],
-      description: response['items'][0]['snippet']['description']
-    }
   end
 end
